hmap: initial commit
The hmap is a datastructure that maps keys to sets of objects. It assumes that
the key is embedded in the object, which is a common pattern in our code. It
is efficient in CPU time, cache usage, and total memory consumption.

The implementation is an open-addressed hashtable that uses Robin Hood hashing
and splits the hash and value arrays for better cache utilization.